Differentiate between ID types in UoW's identity map
Previous use of document IDs as array indexes lead to ambiguity when handling numeric strings. Serializing values to respect ID types will avoid this, although it may be too strict for floats and integers, which are comparable in MongoDB.

Additionally, this change requires that numeric IDs be fetched by the appropriate type, as shown in IdTest::testCollectionId().